refactor: moved all external use-cases inside `application.external`
